requirements.txt
Flask>=3.0.3
requests>=2.32.0
python-dotenv>=1.0.0
Pillow>=10.0.0
streamlit>=1.28.0

